requirements.txt
asgiref==3.4.1
Django==3.2.12
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
